Feature/enable graph levels (#32)
* Set up ability to change levels

* Set up levels to call backedn for lineage with levels

* Removed levels capabilities
